Renamed files for `JoltSliderJointImpl3D`
